feat: pet feature for pet owner

+++ backend/model/jwt_claim.go
@@ -4,5 +4,5 @@ import "github.com/golang-jwt/jwt/v5"
 
 type JWTClaim struct {
 	jwt.RegisteredClaims
-	UserID int64 `json:"userId,omitempty"`
+	UserID uint `json:"userId,omitempty"`
 }
